feat: FE create machine
