+++ frontend/src/MachineCard.css
@@ -1,4 +1,8 @@
 #machine-card {
   border: 1px solid greenyellow;
-  padding: 0 2rem ;
+  padding: 0 2rem;
+}
+
+#machine-card:hover {
+  border: 1px solid red;
 }
