+++ frontend/src/Machines.css
@@ -1,7 +1,7 @@
 #createMachine {
   border: 1px solid gray;
   padding: 0.5rem;
-  margin: 0;
+  margin: auto;
 
   display: flex;
   flex-direction: column;
@@ -10,7 +10,7 @@
   gap: 0.5rem
 }
 
-#createMachine h2{
+#createMachine h2 {
   text-align: center;
   margin: 0;
 }
@@ -20,4 +20,15 @@
   flex-direction: row-reverse;
   justify-content: space-between;
   flex-grow: 1;
+}
+
+#machinesBos {
+  display: flex;
+  flex-direction: column;
+}
+
+#content {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
 }
