+++ frontend/src/constants.ts
@@ -1,0 +1,5 @@
+export const SENSOR_TYPES = ["HF+", "TcAg", "TcAs"]
+
+export const MACHINE_TYPES = ["Fan", "Pump"]
+
+export const SENSOR_MAP = [["TcAg", "TcAs", "HF+"], ["HF+"]]
